// src/@types/theme/syled.d.ts
export interface DefaultTheme {
  theme: 'dark' | 'light'
  palette: {
    bfscompany: {
      primary: string
      secondary: string
      tertiary: string
      accent: string
      soft: {
        primary: string
        secondary: string
        tertiary: string
      }
    }
    company: {
      primary: string
      secondary: string
      tertiary: string
      accent: string
      soft: {
        primary: string
        secondary: string
        tertiary: string
      }
    }
    background: {
      paper: string
    }
    primary: {
      background: {
        default: string
        dark: string
        light: string
        accent: string
      }
      color: {
        default: string
        dark: string
        light: string
        accent: string
      }
    }
    secondary: {
      background: {
        default: string
        dark: string
        light: string
        accent: string
      }
      color: {
        default: string
        dark: string
        light: string
        accent: string
      }
    }
    tertiary: {
      background: {
        default: string
        dark: string
        light: string
        accent: string
      }
      color: {
        default: string
        dark: string
        light: string
        accent: string
      }
    }
    quaternary: {
      background: {
        default: string
        dark: string
        light: string
        accent: string
      }
      color: {
        default: string
        dark: string
        light: string
        accent: string
      }
    }
  }
}